+++ preferences/src/config.tsx
@@ -31,6 +31,7 @@ export type ChewingTsfConfig = {
   font_highlight_fg_color: string;
   font_highlight_bg_color: string;
   font_number_fg_color: string;
+  cand_list_border_color: string;
   keyboard_layout: number;
   simulate_english_layout: number;
   keybind: [KeybindValue];